requirements.txt
django==5.0.2
djangorestframework==3.14.0
Pillow==10.2.0
drf_yasg==1.21.7
python-dotenv==1.0.1
django-filter==23.5
celery[redis]==5.3.6
redis==5.0.1
djangorestframework-simplejwt==5.3.1
fcm-django==2.0.1